requirements.txt
pbr>=0.5.21,<1.0
pytest>=2.5.2
docopt>=0.6.1
docker-py>=0.2.3
colorama>=0.2.5
